Fixed vibrotactile task data bugs.

+++ MATLAB Functions/OmniTrakFileRead Subfunctions/OmniTrakFileRead_ReadBlock_V1_PRIMARY_MODULE.m
@@ -11,5 +11,5 @@ function data = OmniTrakFileRead_ReadBlock_V1_PRIMARY_MODULE(fid,data)
 % fwrite(fid,str,'uchar');
 
 data = OmniTrakFileRead_Check_Field_Name(data,'module');                    %Call the subfunction to check for existing fieldnames.
-N = fread(fid,1,'uint8');                                                   %Read in the number of characters.
+N = fread(fid,1,'uint16');                                                  %Read in the number of characters.
 data.module(1).name = fread(fid,N,'*char')';                                %Read in the characters of the primary module name.
